huge concurrency commit: + two bot grabbers realisation [dorm - all in one thread, home - all bots execute separated (in own threads)] + heavyweight bot methods (webhooks) execution supported - it means that huge method could be executed in separated thread + for heavyweight webhooks release chat actions or in the framework language - fast actions, fast actions notify user that bot is executing and needs time + gateway detach producers calls - its mean that now all calls to bot async was: bot <-sync- gateway <-sync- producer now: bot <-async- gateway <-sync- producer + rework events system - migrated to spring event factory and separated on two types of events was: onBotRegistry - not all beans who needed bots data ready to receive that event now: 1) OnBotRegistry - the same 2) BotsReady - after context refresh with bot registry service and application context + telegram http client supports async calls (ok http adapter too) + migrates all identifiers from int to long PS some implementations are very snapshot, but working

+++ telegram-bot-api/src/main/java/com/goodboy/telegram/bot/api/PassportFile.java
@@ -46,10 +46,10 @@ public class PassportFile {
     /**
      * File size
      */
-    private Integer fileSize;
+    private Long fileSize;
 
     /**
      * 	Unix time when the file was uploaded
      */
-    private Integer fileData;
+    private Long fileData;
 }
